refactor OperatorMap component to use maps folder in src/app/maps/Map.tsx
