+++ src/app/maps/Map.tsx
@@ -1,22 +1,19 @@
 import React, { useState, useEffect, useRef } from "react";
 import { StyleSheet, View, TouchableOpacity, Text, Alert } from "react-native";
-import MapView, { Marker } from "react-native-maps";
+import MapView, { MapViewProps, Region } from "react-native-maps";
 import * as Location from "expo-location";
 import Icon from "react-native-vector-icons/MaterialIcons";
 import { useTranslation } from "react-i18next";
-import LocationMarker from "../components/LocationMarker";
-import { Order, OrderStatus } from "../types/Order";
-import FirestoreManager from "../services/FirestoreManager";
-import { Button } from "../ui/Button";
+import LocationMarker from "../../components/LocationMarker";
 
-type LocationType = {
-  latitude: number;
-  longitude: number;
-  latitudeDelta: number;
-  longitudeDelta: number;
-};
+const topButtonPadding = 60;
+const sideButtonPadding = 30;
 
-const OperatorMap = ({ navigation }: any) => {
+const MapOverview = ({ navigation }: any) => {
+  type MapViewRef = {
+    animateToRegion: (marker: LocationType, duration?: number) => void;
+  };
+
   const mapRef = useRef<MapView | null>(null);
   const [currentRegion, setCurrentRegion] = useState({
     latitude: 37.789,
@@ -27,9 +24,20 @@ const OperatorMap = ({ navigation }: any) => {
 
   const { t } = useTranslation();
 
+  type LocationType = {
+    latitude: number;
+    longitude: number;
+    latitudeDelta: number;
+    longitudeDelta: number;
+  };
+
   const [marker, setMarker] = useState<LocationType | null>(null);
   const [loading, setLoading] = useState(true);
   const [autoCenter, setAutoCenter] = useState(true);
+
+  function animateToRegion(marker: LocationType, duration: number) {
+    mapRef.current?.animateToRegion(marker, 500);
+  }
 
   useEffect(() => {
     const initMap = async () => {
@@ -47,7 +55,7 @@ const OperatorMap = ({ navigation }: any) => {
 
   useEffect(() => {
     if (autoCenter && marker) {
-      mapRef.current?.animateToRegion(marker, 500);
+      animateToRegion(marker, 500);
     }
   }, [autoCenter, marker]);
 
@@ -83,93 +91,110 @@ const OperatorMap = ({ navigation }: any) => {
   };
 
   const toggleAutoCenter = () => {
-    if (marker) {
-      mapRef.current?.animateToRegion(marker, 1500);
+    if (marker && mapRef.current) {
+      animateToRegion(marker, 1500);
     }
     setTimeout(() => {
       setAutoCenter(true);
     }, 500);
   };
 
-  const [orders, setOrders] = useState<Order[]>([]);
-
-  useEffect(() => {
-    async function fetchOrders() {
-      const firestoreManager = new FirestoreManager();
-      const pendingOrders = await firestoreManager.queryOrder(
-        "status",
-        OrderStatus.Pending
-      );
-      const acceptedOrders = await firestoreManager.queryOrder(
-        "status",
-        OrderStatus.Accepted
-      );
-      if (pendingOrders && acceptedOrders) {
-        setOrders([...pendingOrders, ...acceptedOrders]);
-      }
-    }
-
-    fetchOrders();
-  }, []);
-
   return (
-    <View style={StyleSheet.absoluteFillObject}>
+    <View style={styles.container}>
       <MapView
         onPanDrag={() => {
           setAutoCenter(false);
         }}
         testID="map-view"
         ref={mapRef}
-        style={StyleSheet.absoluteFillObject}
+        style={styles.map}
         initialRegion={currentRegion}
         onRegionChangeComplete={setCurrentRegion}
       >
         {marker && <LocationMarker coordinate={marker} />}
-        {orders.map((order) => (
-          <Marker
-            coordinate={{
-              latitude: order.getUsrLocation().latitude,
-              longitude: order.getUsrLocation().longitude,
-            }}
-            key={order.getId()}
-            title={t(order.getItem().getName() as "items.first-aid")}
-            description={order.getOrderDate().toLocaleString()}
-            pinColor={
-              order.getStatus() === OrderStatus.Pending ? "yellow" : "green"
-            }
-          />
-        ))}
       </MapView>
 
       {loading && (
-        <View className="flex items-center justify-center bg-white/50 w-full h-full">
+        <View style={styles.loadingContainer}>
           <Text>Loading your location...</Text>
         </View>
       )}
 
-      <Button
+      <TouchableOpacity
         testID="my-location-button"
+        style={[styles.button, styles.buttonTopRight]}
         onPress={toggleAutoCenter}
-        style="primary"
-        className="absolute top-12 right-12 w-16 h-16"
       >
-        <Icon name="my-location" size={24} color="#fff" />
-      </Button>
+        <Icon name="my-location" size={24} color="#000" />
+      </TouchableOpacity>
 
-      <Button
+      <TouchableOpacity
+        testID="user-drawer-button"
+        style={[styles.button, styles.buttonTopLeft]}
+        onPress={() => {
+          navigation.toggleDrawer({
+            latitude: currentRegion.latitude,
+            longitude: currentRegion.longitude,
+          });
+        }}
+      >
+        <Icon name="menu" size={24} color="#000" />
+      </TouchableOpacity>
+
+      <TouchableOpacity
         testID="order-button"
+        style={[styles.button, styles.buttonBottomRight]}
         onPress={() => {
           navigation.navigate("OrderMenu", {
             latitude: currentRegion.latitude,
             longitude: currentRegion.longitude,
           });
         }}
-        style="primary"
-        text={t("map.pending-orders-button")}
-        className="absolute bottom-12 right-12 w-48 h-16"
-      />
+      >
+        <Text style={styles.buttonText}>{t("map.order-button")}</Text>
+      </TouchableOpacity>
     </View>
   );
 };
 
-export default OperatorMap;
+const styles = StyleSheet.create({
+  container: {
+    ...StyleSheet.absoluteFillObject,
+  },
+  map: {
+    ...StyleSheet.absoluteFillObject,
+  },
+  button: {
+    position: "absolute",
+    backgroundColor: "rgba(255,255,255, 1)",
+    alignItems: "center",
+    justifyContent: "center",
+    borderRadius: 20,
+    padding: 15,
+  },
+  buttonTopRight: {
+    top: topButtonPadding,
+    right: sideButtonPadding,
+  },
+  buttonBottomRight: {
+    bottom: 40,
+    right: sideButtonPadding,
+  },
+
+  buttonTopLeft: {
+    top: topButtonPadding,
+    left: sideButtonPadding,
+  },
+  buttonText: {
+    fontSize: 20,
+    color: "#000",
+  },
+  loadingContainer: {
+    flex: 1,
+    justifyContent: "center",
+    alignItems: "center",
+    backgroundColor: "rgba(255,255,255, 0.7)",
+  },
+});
+
+export default MapOverview;
